* further cleanup patches from andre (Mathematics.h): 	- The code fails badly in case size==1. 	-Also, clean up the placement of spaces of that function according 	 to README.developers. 	 The code was broken for n<=0: For n==0 it would print only 	 "name=[", and it would segfault for n<0. Fix that up as well. 	 We're compiling with -O, so gcc will unroll the last iteration of 	 the loop, i.e. the new code should not be slower than the old code. 	- Kill four pointless loops and don't segfault for rows<=0 or cols<=0. * remove broken KerthiPrimalSVM * implement CSparseLinearClassifier * derive CSVMLin from CSparseLinearClassifier * add example on how to use SVMLin from python * update AUTHORS, CONTRIBUTIONS * make all the cache sizes in all the feature objects Constructors 	INT (as they are in MB anyway)
Yes. We now have Sindhwani/Keerthis primal svm in shogun.

+++ src/features/FKFeatures.cpp
@@ -13,7 +13,7 @@
 #include "features/StringFeatures.h"
 #include "lib/io.h"
 
-CFKFeatures::CFKFeatures(LONG size, CHMM* p, CHMM* n) : CRealFeatures(size)
+CFKFeatures::CFKFeatures(INT size, CHMM* p, CHMM* n) : CRealFeatures(size)
 {
 	pos_prob=NULL;
 	neg_prob=NULL;
